requirements.txt
APScheduler==3.10.1
colorama==0.4.6
python-telegram-bot==20.2
SQLAlchemy==2.0.7
